+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.4 (WIP)
 
 * Make login form images configurable
+* Make the `Permissions` field inherit from `Field.Schema` & allow adding custom fields
 
 ## 0.8.3 (2022-11-02)